+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 1.3.1
+
+Add a `scheduleTask` callback via constructor or via `trigger()` to specify a own
+task scheduler method, e.g. `Future.microtask`
+
 ## 1.2.0
 
 Add a `isTest` constructor parameter + `executeNow()` method control execution